pyproject.toml
[tool.poetry]
name = "test-task-paritet"
version = "0.1.0"
description = ""
authors = ["SiberianFalcon <bq.Sokol@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
drf-extra-fields = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
